Files modified: 1
Modified 1 file:geom/svg/path/utils/SVGPathUtils.swift

+++ geom/svg/path/utils/SVGPathUtils.swift
@@ -43,7 +43,7 @@ class SVGPathUtils {
                     Swift.print("curveTo")
                     pos += CGPoint(params[i+4],params[i+5]);
                     let controlP1:CGPoint = isLowerCase ? CGPoint(prevP.x + params[i],prevP.y+params[i+1]) : CGPoint(params[i],params[i+1]);
-                    prevC = isLowerCase ? CGPoint(prevP.x+CGFloat(params[i+2]),prevP.y+CGFloat(params[i+3])) : CGPoint(params[i+2],params[i+3]);/*aka controlP2*/
+                    prevC = isLowerCase ? CGPoint(prevP.x + params[i+2],prevP.y+params[i+3]) : CGPoint(params[i+2],params[i+3]);/*aka controlP2*/
                     CGPathAddCurveToPoint(path, nil, prevC.x, prevC.y, controlP1.x, controlP1.y, pos.x, pos.y)//CubicCurveModifier.cubicCurveTo(graphics, prevP, controlP1, prevC, pos);
                     i += 6;
                     break;
